requirements.txt
python-edlib==1.3.*
pysam==0.22.*